mir: Split vtable.rs
