+++ lib/skc_mir/src/vtable.rs
@@ -1,9 +1,7 @@
-use crate::library::LibraryExports;
 use serde::{Deserialize, Serialize};
-use shiika_core::{names::*, ty::*};
-use skc_hir::{SkClass, SkType, SkTypes};
+use shiika_core::names::*;
+use skc_hir::SkClass;
 use std::collections::HashMap;
-use std::collections::VecDeque;
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 pub struct VTable {
@@ -64,71 +62,5 @@ impl VTable {
     /// Returns the list of method names, ordered by the index.
     pub fn to_vec(&self) -> &Vec<MethodFullname> {
         &self.fullnames
-    }
-}
-
-#[derive(Clone, Debug, Default, Serialize, Deserialize)]
-pub struct VTables {
-    // REFACTOR: how about just use `type`
-    vtables: HashMap<ClassFullname, VTable>,
-}
-
-impl VTables {
-    /// Build vtables of the classes
-    pub fn build(sk_types: &SkTypes, imports: &LibraryExports) -> VTables {
-        let mut vtables = HashMap::new();
-        let mut queue = sk_types.keys().cloned().collect::<VecDeque<_>>();
-        let null_vtable = VTable::null();
-        while !queue.is_empty() {
-            let name = queue.pop_front().unwrap();
-            // Check if already processed
-            if vtables.contains_key(&name) || imports.sk_types.contains_key(&name) {
-                continue;
-            }
-
-            let sk_type = sk_types
-                .get(&name)
-                .unwrap_or_else(|| panic!("class not found: {}", name));
-            if let SkType::Class(class) = sk_type {
-                let super_vtable;
-                if let Some(superclass) = &class.superclass {
-                    let super_name = superclass.base_fullname();
-                    if let Some(x) = vtables.get(&super_name) {
-                        super_vtable = x;
-                    } else if let Some(x) = imports.vtables.vtables.get(&super_name) {
-                        super_vtable = x;
-                    } else {
-                        queue.push_front(super_name);
-                        queue.push_back(class.base.fullname());
-                        continue;
-                    }
-                } else {
-                    // The class Object does not have a superclass.
-                    super_vtable = &null_vtable;
-                }
-                let vtable = VTable::build(super_vtable, class);
-                vtables.insert(class.base.fullname(), vtable);
-            }
-        }
-        VTables { vtables }
-    }
-
-    /// Return the index of the method when invoking it on the object
-    pub fn method_idx(
-        &self,
-        obj_ty: &TermTy,
-        method_name: &MethodFirstname,
-    ) -> Option<(&usize, usize)> {
-        self.vtables.get(&obj_ty.vtable_name()).map(|vtable| {
-            let idx = vtable
-                .get(method_name)
-                .unwrap_or_else(|| panic!("[BUG] `{}' not found in {}", &method_name, &obj_ty));
-            (idx, vtable.size())
-        })
-    }
-
-    /// Returns iterator over each vtable
-    pub fn iter(&self) -> std::collections::hash_map::Iter<'_, ClassFullname, VTable> {
-        self.vtables.iter()
     }
 }
